+++ tests/testthat/test-req-cookies.R
@@ -3,7 +3,7 @@ test_that("can read/write cookies", {
 
   set_cookie <- function(req, name, value) {
     request_test("/cookies/set/:name/:value", name = name, vaue = value) %>%
-      req_cookie_file(cookie_path) %>%
+      req_cookie_preserve(cookie_path) %>%
       req_perform()
   }
   set_cookie(req, "x", "a")
@@ -13,7 +13,7 @@ test_that("can read/write cookies", {
   expect_snapshot(readLines(cookie_path)[-(1:4)])
 
   cookies <- request_test("/cookies") %>%
-    req_cookie_file(cookie_path) %>%
+    req_cookie_preserve(cookie_path) %>%
     req_perform() %>%
     resp_body_json() %>%
     .$cookies